Split rpc methods into different plugins
