+++ sequencer/src/main/java/net/consensys/linea/rpc/services/LineaSetExtraDataEndpointPlugin.java
@@ -13,30 +13,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.rpc;
-
-import static net.consensys.linea.sequencer.modulelimit.ModuleLineCountValidator.createLimitModules;
+package net.consensys.linea.rpc.services;
 
 import com.google.auto.service.AutoService;
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.AbstractLineaRequiredPlugin;
 import net.consensys.linea.extradata.LineaExtraDataHandler;
+import net.consensys.linea.rpc.methods.LineaSetExtraData;
 import org.hyperledger.besu.plugin.BesuContext;
 import org.hyperledger.besu.plugin.BesuPlugin;
-import org.hyperledger.besu.plugin.services.BesuConfiguration;
-import org.hyperledger.besu.plugin.services.BlockchainService;
 import org.hyperledger.besu.plugin.services.RpcEndpointService;
-import org.hyperledger.besu.plugin.services.TransactionSimulationService;
 
 /** Registers RPC endpoints. This class provides RPC endpoints under the 'linea' namespace. */
 @AutoService(BesuPlugin.class)
 @Slf4j
-public class LineaEndpointServicePlugin extends AbstractLineaRequiredPlugin {
-  private BesuConfiguration besuConfiguration;
+public class LineaSetExtraDataEndpointPlugin extends AbstractLineaRequiredPlugin {
   private RpcEndpointService rpcEndpointService;
-  private TransactionSimulationService transactionSimulationService;
-  private BlockchainService blockchainService;
-  private LineaEstimateGas lineaEstimateGasMethod;
   private LineaSetExtraData lineaSetExtraDataMethod;
 
   /**
@@ -46,13 +38,6 @@ public class LineaEndpointServicePlugin extends AbstractLineaRequiredPlugin {
    */
   @Override
   public void doRegister(final BesuContext context) {
-    besuConfiguration =
-        context
-            .getService(BesuConfiguration.class)
-            .orElseThrow(
-                () ->
-                    new RuntimeException(
-                        "Failed to obtain BesuConfiguration from the BesuContext."));
 
     rpcEndpointService =
         context
@@ -61,30 +46,6 @@ public class LineaEndpointServicePlugin extends AbstractLineaRequiredPlugin {
                 () ->
                     new RuntimeException(
                         "Failed to obtain RpcEndpointService from the BesuContext."));
-
-    transactionSimulationService =
-        context
-            .getService(TransactionSimulationService.class)
-            .orElseThrow(
-                () ->
-                    new RuntimeException(
-                        "Failed to obtain TransactionSimulatorService from the BesuContext."));
-
-    blockchainService =
-        context
-            .getService(BlockchainService.class)
-            .orElseThrow(
-                () ->
-                    new RuntimeException(
-                        "Failed to obtain BlockchainService from the BesuContext."));
-
-    lineaEstimateGasMethod =
-        new LineaEstimateGas(besuConfiguration, transactionSimulationService, blockchainService);
-
-    rpcEndpointService.registerRPCEndpoint(
-        lineaEstimateGasMethod.getNamespace(),
-        lineaEstimateGasMethod.getName(),
-        lineaEstimateGasMethod::execute);
 
     lineaSetExtraDataMethod = new LineaSetExtraData(rpcEndpointService);
 
@@ -97,12 +58,6 @@ public class LineaEndpointServicePlugin extends AbstractLineaRequiredPlugin {
   @Override
   public void beforeExternalServices() {
     super.beforeExternalServices();
-    lineaEstimateGasMethod.init(
-        rpcConfiguration(),
-        transactionPoolValidatorConfiguration(),
-        profitabilityConfiguration(),
-        createLimitModules(tracerConfiguration()),
-        l1L2BridgeSharedConfiguration());
     lineaSetExtraDataMethod.init(
         new LineaExtraDataHandler(rpcEndpointService, profitabilityConfiguration()));
   }
